Main code change: Enable sparse vector computation when calculating value/gradient/hessian-vector
In previous versions, we assume the model coefficients are represented as a dense vector, and all the computations are done in a dense way. This RB will remove this assumption and enable sparse vector computation if the model coefficients are represented as a sparse vector.

This RB also contains a few other minor changes:
1. Use kryo as the default serializer.
2. Address some of the code style issues.