Refactor the relationship between Algorithms and Optimizers
This change shifts responsibility of creating optimizers to
GeneralLinearAlgorithm, and allows for much more flexibility in defining
new algorithms.

The problem is described in more detail here:

https://docs.google.com/a/linkedin.com/document/d/1sVX7z_yh5qV36636yr7Fkom0o9s1GqHvzxMVkE9Y9hc/edit?usp=sharing

NOTE: This does not implement the proposed solution there. I started on
that path, but it became a much larger project. In order to get going
with the Game migration, I'm making this less invasive fix.

https://jira01.corp.linkedin.com:8443/browse/ML-1023

RB=670039
G=photon-reviewers
R=yizhou,abain,xazhang,dpeng,pogilvie,bchen,nkatariy
A=xazhang,ashelkov


Former-commit-id: a51bbafb701aa0da6979274254f43c93cae279f1