Fix bug where feature name substrings collide when parsing box constraints
This code previously had the following issue:

If the user specified box constraints where some of the feature keys
were substrings of other feature keys, we would fail here with an
IllegalArgumentException, because we find features in featureKeyToIdMap
based on a substring match. E.g.:

[{"name":"2","term":"*","lowerBound":0}, {"name":"21","term":"*","lowerBound":0}]

The fix is to include the name/term delimiter in the search string.