Improve PhotonLogger stability when jobs are running for hours
Currently PhotonLogger is maintaining a remote output stream with HDFS directly,
the connection could be easily interrupted due to cluster unstability if the job
is running for a long time.

This commit is changing the behavior by firstly writing the logs onto a local
tmp file and then finally copy it from local to HDFS when the logger is closed.