Allow PhotonML to accept LibSVM file format in addition to Avro
Introducing InputDataFormat trait to ease the data format assumption of PhotonML.

After this PR, supposedly we should be able to load LibSVM for
training/validation and also bring in the potential of supporting any other
arbirary format.

I tried to do a minimal refractoring possible for Driver. The main goal is to
decouple the linkage between GLMSuite and core Driver functions; so that now
Driver only cares about an InputDataFormat implementation which will provide
loaded label points and index map if asked, but it won't really bother asking
the details of the implementation.

Ideally, some functionalities in GLMSuite should be further decoupled and
refactored to better locations. But to limit the scope, we are not addressing
those issues here.