Fix scalability issue with DefaultIndexMap.size
Previously there was no implementation for "DefaultIndexMap.size", so
the runtime fell back to the "size" implementation in TraversableOnce,
which iterates over the collections of entries each time. Since the
backing map "featureNameToIdMap" is immutable, we can assume that its
size never changes, and cache the size when it's passed in.