Issue#156: allowing randomEffectId fetching in metadatMap as an option
This change will allow randomEffectId fetching to fall back to metadataMap if
a key is not present in the Avro field directly. This will make the Avro schema
a bit more flexible in the sense that users won't necessarily to change Avro
schema definition when changing the set of ids provided for training.