Update GAME DriverTest for DataReader changes
Note: The training data had to be updated because the databricks Avro
reading code is stricter about enforcing types. This is because the data
must be strongly typed to fit into a DataFrame schema. We may have to
help end users through fixing overly-permissive union types in their
data schema, but this should be a relatively quick fix for any
reasonable dataset.