Reimplement AvroDataReader with a direct-to-avro approach rather than using spark-avro
The original PR #179 used spark-avro for reading Avro files, which had two problems:

1) It used a lot more memory on the grid (cause for this still not identified)

2) It caused a lot of shuffling, which made data reading much slower. The shuffling happened because the pattern for
Spark's DataFrame typically hides partition counts from the user, treating this as an implementation detail that a
consumer shouldn't need to worry about at the DataFrame level of abstraction. This is usually true, but in the case of
Photon, we still use partition counts in a few places to tune the performance of algorithms. The DataFrame runtime has
no knowledge of this and badly underestimates the number of partitions needed, which causes a huge repartition/shuffle
step during reading. Here we pass the partition count explicitly to the "read" function to get around this problem. Once
we revisit Photon's sensitivity to choice of partition count, we can remove this parameter.