Fix issue where IndexMapProjector incorrectly projects dense vectors
The idea behind building this projecting mapping is that we can reduce feature dimensionality for a RE problem if there
are features in the full space that are all zeros in the RE sub-problem. The previous method used "activeKeysIterator" for
this, but for DenseVector, this method just returns all indices. Hence reduction was never happening.