Fix issue where IndexMapProjector produces nondeterministic results
Previously we were scanning only the active data to determine the nonzero indices of a RE dataset. This can produce
nondeterministic results when we recompute and resample the active data, because the dimensionality of the feature
vectors can change and get out of sync with the model.

The fix is to scan both active and passive data to determine the feature projection.