Change hyperparameter candidate generation to a Sobol generator
Sobol generators produce equidistant points, which cover the space better than a simple uniform distribution. With this
change, the EI acquisition function appears less likely to get stuck in local optima in experiments. Hence we also
switch to EI prediction transformation so that there's no need to specify an exploration parameter as with UCB.